refac: moved update ledger default
